int->double fix

+++ imdbdrinks_ratingsmodule/Repositories/DatabaseRatingRepository.cs
@@ -88,7 +88,7 @@ class DatabaseRatingRepository : IRatingRepository
                     RatingId = reader.GetInt32(reader.GetOrdinal("RatingId")),
                     ProductId = reader.GetInt32(reader.GetOrdinal("ProductId")),
                     UserId = reader.GetInt32(reader.GetOrdinal("UserId")),
-                    RatingValue = reader.GetInt32(reader.GetOrdinal("RatingValue")),
+                    RatingValue = reader.GetDouble(reader.GetOrdinal("RatingValue")),
                     RatingDate = reader.GetDateTime(reader.GetOrdinal("RatingDate")),
                     IsActive = reader.GetByte(reader.GetOrdinal("IsActive")) == 1,
                 };
